# laravel-dashboard/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Scraper API",
        "description": "API for managing companies and job postings scraped from LinkedIn",
        "contact": {
            "email": "admin@jobscraper.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/companies/exists": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Check if a company exists by name",
                "operationId": "c6b044c5acbed700374f5ea7a1d2f59b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Example Company Ltd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company existence check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "exists": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "company": {
                                            "properties": {
                                                "company_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Example Company Ltd"
                                                },
                                                "vat": {
                                                    "type": "string",
                                                    "example": "NL123456789B01",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Active",
                                                    "nullable": true
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Business Street",
                                                    "nullable": true
                                                },
                                                "zipcode": {
                                                    "type": "string",
                                                    "example": "1234AB",
                                                    "nullable": true
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Amsterdam",
                                                    "nullable": true
                                                },
                                                "website": {
                                                    "type": "string",
                                                    "example": "https://example.com",
                                                    "nullable": true
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "info@example.com",
                                                    "nullable": true
                                                },
                                                "employees": {
                                                    "type": "integer",
                                                    "example": 50,
                                                    "nullable": true
                                                },
                                                "industrycode": {
                                                    "type": "string",
                                                    "example": "6201",
                                                    "nullable": true
                                                },
                                                "industrydesc": {
                                                    "type": "string",
                                                    "example": "Computer programming activities",
                                                    "nullable": true
                                                },
                                                "companytype": {
                                                    "type": "string",
                                                    "example": "BV",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key for authentication",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}